Refactor transforms into their own write file
